backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
# python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
PyPDF2==3.0.1
openai>=1.0.0
python-dotenv>=1.0.0
pathlib>=1.0.0
mammoclip>=0.1.8
torch>=2.0.0
Pillow>=9.0.0
openai>=1.0.0
tiktoken>=0.5.0
cachetools>=5.0.0
python-dateutil>=2.8.2